Convert CompactSortedFilesIT tests to unit tests
